+++ src/commands/explorerArrows.ts
@@ -1,11 +1,11 @@
 import * as vscode from 'vscode';
-import * as helpers from './../helpers';
-import * as versioning from './../helpers/versioning';
-import * as i18n from './../i18n';
-import * as outdatedMessage from './../messages/outdated';
+import * as helpers from '../helpers';
+import * as versioning from '../helpers/versioning';
+import * as i18n from '../i18n';
+import * as outdatedMessage from '../messages/outdated';
 
-/** Command to toggle the folder icons. */
-export const toggleFolderArrows = async () => {
+/** Command to toggle the explorer arrows. */
+export const toggleExplorerArrows = async () => {
     if (!versioning.checkVersionSupport('1.18.0-insider')) {
         outdatedMessage.showOutdatedMessage();
         return Promise.reject('Outdated version of vscode!');
@@ -19,21 +19,21 @@ export const toggleFolderArrows = async () => {
     }
 };
 
-/** Show QuickPick items to select preferred configuration for the folder icons. */
+/** Show QuickPick items to select preferred configuration for the explorer arrows. */
 const showQuickPickItems = (status: boolean) => {
     const on: vscode.QuickPickItem = {
         description: i18n.translate('toggleSwitch.on'),
-        detail: i18n.translate(`hidesExplorerArrows.enableArrows`),
+        detail: i18n.translate(`explorerArrows.enable`),
         label: !status ? '\u2714' : '\u25FB'
     };
     const off: vscode.QuickPickItem = {
         description: i18n.translate('toggleSwitch.off'),
-        detail: i18n.translate(`hidesExplorerArrows.disableArrows`),
+        detail: i18n.translate(`explorerArrows.disable`),
         label: status ? '\u2714' : '\u25FB'
     };
     return vscode.window.showQuickPick(
         [on, off], {
-            placeHolder: i18n.translate('hidesExplorerArrows.toggleArrows'),
+            placeHolder: i18n.translate('explorerArrows.toggle'),
             ignoreFocusOut: false,
             matchOnDescription: true
         });