+++ src/messages/welcome.ts
@@ -1,4 +1,4 @@
-import * as opn from 'opn';
+import * as open from 'open';
 import * as vscode from 'vscode';
 import { activateIcons } from '../commands/activate';
 import * as helpers from './../helpers';
@@ -29,7 +29,7 @@ const handleWelcomeMessageActions = (value) => {
             break;
 
         case i18n.translate('howToActivate'):
-            opn('https://code.visualstudio.com/blogs/2016/09/08/icon-themes#_file-icon-themes');
+            open('https://code.visualstudio.com/blogs/2016/09/08/icon-themes#_file-icon-themes');
             break;
 
         case i18n.translate('neverShowAgain'):